Move OnBeforeRequest* types into the api client middleware module
The `OnBeforeRequestHandler` / `OnBeforeRequestHandlerConfig` types lived in
`metabase/plugins/oss/api` but describe the request-manipulation middleware,
which lives under `metabase/api/client`. Relocate the type definitions to
`metabase/api/client/middleware` and re-export them from the `metabase/api/client`
barrel, so the middleware contract is fully contained in its own module. Pure
move plus import updates; no behavior change.
