+++ frontend/src/metabase/api/client/index.ts
@@ -1,5 +1,9 @@
 export * from "./client";
 export * from "./method";
 export * from "./errors";
+export type {
+  OnBeforeRequestHandler,
+  OnBeforeRequestHandlerConfig,
+} from "./middleware";
 export type { RequestOptions } from "./types";
 export { setLocaleHeader } from "./locale";
