+++ frontend/src/metabase/embedding/lib/override-requests-for-embeds.ts
@@ -1,12 +1,15 @@
 import { sessionPropertiesPath } from "metabase/api";
-import { type RequestMethod, api } from "metabase/api/client";
+import {
+  type OnBeforeRequestHandlerConfig,
+  type RequestMethod,
+  api,
+} from "metabase/api/client";
 import { isEmbedPreview } from "metabase/embedding/config";
 import {
   PLUGIN_API,
   PLUGIN_CONTENT_TRANSLATION,
   PLUGIN_EMBEDDING_SDK,
 } from "metabase/plugins";
-import type { OnBeforeRequestHandlerConfig } from "metabase/plugins/oss/api";
 import type { CardId, DashboardId, ParameterId } from "metabase-types/api";
 
 type EmbedType = "guest" | "static" | "public";
